Core/Misc: Really fixed nopch MSVC build
