+++ src/server/game/Movement/MovementGenerators/WaypointMovementGenerator.h
@@ -28,24 +28,11 @@
 
 #include "DBCStructure.h"
 #include "MovementGenerator.h"
+#include "PathMovementBase.h"
 #include "Timer.h"
 
 class Creature;
 struct WaypointPath;
-
-template<class Entity, class BasePath>
-class PathMovementBase
-{
-    public:
-        PathMovementBase() : _path(), _currentNode(0) { }
-        virtual ~PathMovementBase() { };
-
-        uint32 GetCurrentNode() const { return _currentNode; }
-
-    protected:
-        BasePath _path;
-        uint32 _currentNode;
-};
 
 template<class T>
 class WaypointMovementGenerator;
